+++ src/.env
@@ -1,3 +1,3 @@
-NODE_ENV=prod
-APP_VERSION="1.1.1"
+NODE_ENV=dev
+APP_VERSION=1.1.1
 RELEASE_NUMBER=14
