+++ src/context/SuperAdminHomePageContext/index.ts
@@ -2,3 +2,5 @@ export {
   SuperAdminHomePageContextProvider,
   useSuperAdminHomePageFunction,
 } from "./SuperAdminHomePageContext";
+
+export { useSuperAdminHomePageContext } from "./useSuperAdminHomePageContext";
