+++ src/screens/SuperAdminHomePage/index.ts
@@ -1,0 +1,1 @@
+export { SuperAdminHomePage } from "./SuperAdminHomePage";
